% Auto-generated: do not edit by hand
\name{''OpenfinContextSubscriber}

\alias{''OpenfinContextSubscriber}

\title{OpenfinContextSubscriber component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''OpenfinContextSubscriber(id=NULL, contextTopic=NULL, currentContext=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{contextTopic}{Character. The value for the OpenFin topic to publish values onto.}

\item{currentContext}{Character. The context passed back out when the OpenFin subscription gets called back.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

